backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
graphene_django
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
numpy
pyaudio
webrtcvad
faster_whisper
torch
soundfile
scipy
transformers
sendgrid
Pillow
openai-whisper
<<<<<<< HEAD
django-extensions
=======
pydub
fastapi
langchain
langchain_chroma
langchain_community
langchain_openai
librosa
gtts
pygame
pyaudio
sequence_align
transformers
nltk
torch
aiofiles
>>>>>>> 8c7561de3bbe52f887c9065f39ea052265c0fae5
